Move withdraw test back with the other x domain config rename package and cleanup
