+++ package.json
@@ -14,7 +14,7 @@
     "clean": "wsrun -p $(yarn --silent run pkgparse) -r --fast-exit --parallel --exclude-missing clean",
     "clean_modules": "rm -rf node_modules && find ./packages -type d -maxdepth 2 -name \"node_modules\" -exec rm -r {} +",
     "test": "wsrun -p $(yarn --silent run pkgparse) --fast-exit --no-prefix --exclude-missing --timeout 20000 test",
-    "test:integration-tests": "NODE_ENV=local wsrun -p @eth-optimism*/integration-tests --fast-exit --no-prefix --exclude-missing --timeout 20000 test",
+    "test:x-domain": "NODE_ENV=local wsrun -p @eth-optimism*/x-domain --fast-exit --no-prefix --exclude-missing --timeout 20000 test",
     "test:tx-ingestion": "NODE_ENV=local wsrun -p @eth-optimism*/tx-ingestion --fast-exit --no-prefix --exclude-missing --timeout 20000 test",
     "build": "wsrun -p $(yarn --silent run pkgparse) -r --fast-exit --stages --exclude-missing build",
     "all": "yarn clean && yarn build && yarn test && yarn fix && yarn lint",
