+++ packages/x-domain/tsconfig.json
